♻️ refactor(header): move project related components to its own folder
- relocate `header-project-tkcapturebook.tsx`, `header-project-tkcapturebook-privacy.tsx`, and `header-project-tkcapturebook-terms.tsx` into `project` subfolder
- update import paths in `header.tsx`
